chore: relocate data files
