+++ src/MakaMek.Core/Data/Game/Commands/CommandValidationResult.cs
@@ -1,6 +1,4 @@
-using Sanet.MakaMek.Core.Data.Game.Commands.Server;
-
-namespace Sanet.MakaMek.Core.Models.Game;
+namespace Sanet.MakaMek.Core.Data.Game.Commands;
 
 public readonly record struct CommandValidationResult(bool IsValid, ErrorCode? ErrorCode = null)
 {
