+++ tests/MakaMek.Core.Tests/Data/Game/Commands/Server/ErrorCommandTest.cs
@@ -1,4 +1,5 @@
 ﻿using NSubstitute;
+using Sanet.MakaMek.Core.Data.Game.Commands;
 using Sanet.MakaMek.Core.Data.Game.Commands.Server;
 using Sanet.MakaMek.Core.Models.Game;
 using Sanet.MakaMek.Core.Services.Localization;
